Adjusting file structure to keep category selector at the top
